Move categories functions to separate file
