+++ registry/src/extensions.rs
@@ -1,7 +1,6 @@
 use crate::download::latest_version;
 use crate::errors::ExtensionRegistryError;
 use actix_web::web::Data;
-use log::{debug, info};
 use serde_json::{json, Value};
 use sqlx::{Pool, Postgres};
 
@@ -99,129 +98,4 @@ pub async fn get_extension_id(
         .fetch_one(&mut tx)
         .await?;
     Ok(id.id)
-}
-
-// Update categories for a given extension
-pub async fn update_extension_categories(
-    extension_id: i64,
-    categories: Vec<String>,
-    conn: Data<Pool<Postgres>>,
-) -> Result<(), ExtensionRegistryError> {
-    debug!("categories: {:?}", categories);
-    let mut tx = conn.begin().await?;
-    // Get category IDs
-    let new_ids = get_category_ids(categories, conn).await?;
-    // Get existing categories for the extension
-    let existing = sqlx::query!(
-        "SELECT category_id FROM extensions_categories WHERE extension_id = $1;",
-        extension_id as i32
-    )
-    .fetch_all(&mut tx)
-    .await?;
-
-    let existing_ids: Vec<i64> = existing
-        .into_iter()
-        .map(|x| x.category_id.unwrap() as i64)
-        .collect();
-
-    // If id not in existing, add
-    for category_id in new_ids.clone() {
-        if !existing_ids.contains(&category_id) {
-            info!(
-                "adding category_id {} to extension_id {}",
-                category_id, extension_id
-            );
-            sqlx::query!(
-                "
-                INSERT INTO extensions_categories(extension_id, category_id)
-                VALUES ($1, $2)
-                ",
-                extension_id as i32,
-                category_id as i32
-            )
-            .execute(&mut tx)
-            .await?;
-        }
-    }
-    // If existing not in ids, delete
-    for category_id in existing_ids {
-        if !new_ids.contains(&category_id) {
-            info!(
-                "removing category_id {} from extension_id {}",
-                category_id, extension_id
-            );
-            sqlx::query!(
-                "
-                DELETE FROM extensions_categories
-                WHERE extension_id = $1
-                AND category_id = $2
-                ",
-                extension_id as i32,
-                category_id as i32
-            )
-            .execute(&mut tx)
-            .await?;
-        }
-    }
-
-    tx.commit().await?;
-    Ok(())
-}
-
-pub async fn get_category_ids(
-    categories: Vec<String>,
-    conn: Data<Pool<Postgres>>,
-) -> Result<Vec<i64>, ExtensionRegistryError> {
-    let mut ids: Vec<i64> = Vec::new();
-    let mut tx = conn.begin().await?;
-    for slug in categories {
-        let id = sqlx::query!("SELECT id FROM categories WHERE slug = $1", slug)
-            .fetch_optional(&mut tx)
-            .await?;
-        if id.is_some() {
-            ids.push(id.unwrap().id)
-        }
-    }
-    Ok(ids)
-}
-
-pub async fn get_category_names(
-    category_ids: Vec<i64>,
-    conn: Data<Pool<Postgres>>,
-) -> Result<Vec<String>, ExtensionRegistryError> {
-    let mut categories: Vec<String> = Vec::new();
-    let mut tx = conn.begin().await?;
-    for id in category_ids {
-        let name = sqlx::query!("SELECT name FROM categories WHERE id = $1", id)
-            .fetch_one(&mut tx)
-            .await?;
-        categories.push(name.name.unwrap())
-    }
-    Ok(categories)
-}
-
-pub async fn get_categories(
-    extension_id: i64,
-    conn: Data<Pool<Postgres>>,
-) -> Result<Vec<String>, ExtensionRegistryError> {
-    // let mut category_ids: Vec<i64> = Vec::new();
-    let mut categories: Vec<String> = Vec::new();
-    let mut tx = conn.begin().await?;
-    let category_ids = sqlx::query!(
-        "SELECT category_id FROM extensions_categories WHERE extension_id = $1",
-        extension_id as i32
-    )
-    .fetch_all(&mut tx)
-    .await?;
-    let category_ids: Vec<i64> = category_ids
-        .into_iter()
-        .map(|x| x.category_id.unwrap() as i64)
-        .collect();
-    for id in category_ids {
-        let name = sqlx::query!("SELECT name FROM categories WHERE id = $1", id as i32)
-            .fetch_one(&mut tx)
-            .await?;
-        categories.push(name.name.unwrap())
-    }
-    Ok(categories)
 }
