+++ registry/src/lib.rs
@@ -1,6 +1,7 @@
 use sqlx::postgres::{PgConnectOptions, PgPoolOptions};
 use sqlx::{ConnectOptions, Pool, Postgres};
 
+pub mod categories;
 pub mod config;
 pub mod download;
 pub mod errors;
