+++ registry/src/routes/extensions.rs
@@ -1,12 +1,10 @@
 //! Functionality related to publishing a new extension or version of an extension.
 
+use crate::categories::{get_categories, update_extension_categories};
 use crate::config::Config;
 use crate::download::latest_version;
 use crate::errors::ExtensionRegistryError;
-use crate::extensions::{
-    add_extension_owner, check_input, extension_owners, get_categories, latest_license,
-    update_extension_categories,
-};
+use crate::extensions::{add_extension_owner, check_input, extension_owners, latest_license};
 use crate::token::validate_token;
 use crate::uploader::upload_extension;
 use crate::views::extension_publish::ExtensionUpload;
